Move new references panel into codeintel folder
And drop the "cool"
