+++ client/web/src/codeintel/ReferencesPanel.module.scss
